requirements.txt
batchgenerators==0.25
einops==0.6.1
fvcore==0.1.5.post20221221
matplotlib==3.7.1
MedPy==0.4.0
mmengine==0.7.3
monai==1.1.0
numpy==1.24.3
pandas==2.0.1
PyYAML==6.0
PyYAML==6.0
seaborn==0.12.2
SimpleITK==2.2.1
SimpleITK==2.2.1
skimage==0.0
timm==0.6.13
torch==1.12.1.post201
torch==2.0.0
